backend/requirements.txt
# Core Dependencies
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0

# Database
pymongo==4.6.0

# Authentication
PyJWT==2.8.0
bcrypt==4.1.2

# AI/ML Models for Analysis
transformers==4.36.0
torch==2.1.2
selfcheckgpt==0.2.1
sentence-transformers==2.2.2
openai==1.6.1
anthropic==0.7.8

# NLP Processing
spacy==3.7.2
nltk==3.8.1

# Image Processing
Pillow==10.1.0
opencv-python==4.9.0.80

# Utilities
requests==2.31.0
python-multipart==0.0.6
validators==0.22.0
